Update polling api for eager picks
