+++ website/draft_site/drafts/urls.py
@@ -12,5 +12,7 @@ urlpatterns = [
     path('draft/<int:draft_id>/seat/<int:seat>/all-picks', views.all_picks, name='all_picks'),
     path('draft/<int:draft_id>/pick-card', views.pick_card, name='pick_card'),
     path('draft/<int:draft_id>/seat/<int:seat>/waiting-for-drafters', views.api.waiting_for_drafters,
-         name='waiting_for_drafters')
+         name='waiting_for_drafters'),
+    path('draft/<int:draft_id>/seat/<int:seat>/is-pack-available', views.api.is_pack_available,
+         name='is_pack_available')
 ]
