+++ website/draft_site/drafts/views/api/__init__.py
@@ -1,1 +1,2 @@
 from .waiting_for_drafters import waiting_for_drafters
+from .is_pack_available import is_pack_available
